break circular dependency between CommonTools and JetMETCorrections
