+++ CommonTools/ParticleFlow/plugins/BuildFile.xml
@@ -11,7 +11,6 @@
   <use name="FWCore/Framework"/>
   <use name="FWCore/MessageLogger"/>
   <use name="FWCore/ParameterSet"/>
-  <use name="JetMETCorrections/JetCorrector"/>
   <use name="CommonTools/Utils"/>
   <use name="CommonTools/ParticleFlow"/>
   <use name="CalibFormats/HcalObjects"/>
