+++ CommonTools/ParticleFlow/plugins/SealModule.cc
@@ -1,7 +1,6 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
 
 #include "CommonTools/ParticleFlow/plugins/PFMET.h"
-#include "CommonTools/ParticleFlow/plugins/Type1PFMET.h"
 #include "CommonTools/ParticleFlow/plugins/PFPileUp.h"
 #include "CommonTools/ParticleFlow/plugins/PFCandidateFwdPtrCollectionFilter.h"
 #include "CommonTools/ParticleFlow/plugins/PFJetFwdPtrProducer.h"
@@ -10,7 +9,6 @@
 #include "CommonTools/ParticleFlow/plugins/DeltaBetaWeights.h"
 
 DEFINE_FWK_MODULE(PFMET);
-DEFINE_FWK_MODULE(Type1PFMET);
 DEFINE_FWK_MODULE(PFPileUp);
 
 DEFINE_FWK_MODULE(PFCandidateFwdPtrCollectionStringFilter);
