+++ CommonTools/ParticleFlow/python/pfType1MET_cff.py
@@ -1,7 +1,7 @@
 import FWCore.ParameterSet.Config as cms
 
 from CommonTools.ParticleFlow.pfMET_cfi  import *
-from CommonTools.ParticleFlow.pfType1MET_cfi  import *
+from JetMETCorrections.Type1MET.pfType1MET_cfi  import *
 
 pfRawMET = pfMET.clone(alias="pfRawMET")
 #pfRawMET.hfCalibFactor = 1.
